chore: refactor db dir + better log seed.ts
